// ManagmentServer/API/UI/src/Components/WebHooks/WebHookList/__generated__/WebHookCreateServerListFragment.graphql.ts
export type GQL_ServerVariant = "MQTT" | "OPC" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type WebHookCreateServerListFragment$data = {
  readonly servers: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly type: GQL_ServerVariant;
      } | null;
    }> | null;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null;
    };
  } | null;
  readonly " $fragmentType": "WebHookCreateServerListFragment";
};
export type WebHookCreateServerListFragment$key = {
  readonly " $data"?: WebHookCreateServerListFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"WebHookCreateServerListFragment">;
};